backend/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
exceptiongroup==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==21.2.0
importlib_metadata==7.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pluggy==1.4.0
pymongo==4.6.2
pytest==8.1.1
pytest-flask==1.3.0
python-dotenv==1.0.1
scipy==1.12.0
tomli==2.0.1
Werkzeug==3.0.1
zipp==3.18.1


# blinker==1.8.2
# click==8.1.7
# Flask==3.0.3
# Flask-Cors==5.0.0
# itsdangerous==2.2.0
# Jinja2==3.1.4
# MarkupSafe==2.1.5
# Werkzeug==3.0.4
